# server/openapi.yaml
openapi: 3.0.3
info:
  title: WeatherGeoBridge API
  version: 1.0.0
  description: >
    天気取得・通知・地図統合Webアプリ WeatherGeoBridge のREST API。
    内部実装は core/weather_client.py・cache/weather_cache.py を薄くラップするのみで、
    Open-Meteo呼び出しロジックはここで重複実装しない。
servers:
  - url: http://localhost:8787
paths:
...
  /api/weather/forecast:
    get:
      summary: 指定した緯度経度の数日先までの予報を取得する
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number } }
        - { name: lon, in: query, required: true, schema: { type: number } }
        - { name: days, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 16, default: 3 } }
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Forecast" }
        "401":
          description: X-API-Keyが不正または未設定
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ForecastDay:
      type: object
      properties:
        date: { type: string, format: date }
        tempMaxC: { type: number }
        tempMinC: { type: number }
        precipitationMm: { type: number }
        weatherCode: { type: integer }
    Forecast:
      type: object
      properties:
        lat: { type: number }
        lon: { type: number }
        daily:
          type: array
          items: { $ref: "#/components/schemas/ForecastDay" }